requirements.txt
discord.py==2.3.2
google-cloud-compute==1.18.0
